# implement div based on Knuth Algorithm D