ref #5 

On my local machine, the test stats as follows

Master branch:
```
Benchmark_SDiv/large/big           	10000000	       145 ns/op	      48 B/op	       1 allocs/op
Benchmark_SDiv/large/fixedbit      	 1000000	      1169 ns/op	     320 B/op	       7 allocs/op
Benchmark_Div/large/big            	 3000000	       471 ns/op	     176 B/op	       3 allocs/op
Benchmark_Div/large/fixedbit       	 1000000	      1146 ns/op	     320 B/op	       7 allocs/op
Benchmark_Div/small/big            	10000000	       218 ns/op	     128 B/op	       3 allocs/op
Benchmark_Div/small/fixedbit       	100000000	        21.4 ns/op	       0 B/op	       0 allocs/op
```

With Knuth Algorithm D:
```
Benchmark_SDiv/large/big           	10000000	       158 ns/op	      48 B/op	       1 allocs/op
Benchmark_SDiv/large/fixedbit      	 3000000	       424 ns/op	     128 B/op	       3 allocs/op
Benchmark_Div/large/big            	 3000000	       501 ns/op	     176 B/op	       3 allocs/op
Benchmark_Div/large/fixedbit       	 3000000	       395 ns/op	     128 B/op	       3 allocs/op
Benchmark_Div/small/big            	10000000	       227 ns/op	     128 B/op	       3 allocs/op
Benchmark_Div/small/fixedbit       	100000000	        21.6 ns/op	       0 B/op	       0 allocs/op
```

faster 😃